requirements.txt
transformers==4.38.1
lightning==2.2.1
pytorch-lightning==2.2.1
datasets==2.17.1
hydra-core==1.3.2
rouge_score
scipy
evaluate
sacrebleu
bitsandbytes
rich
peft
structlog
codetiming
wandb